// packages/sites-scripts/src/common/src/template/types.ts
export type Manifest = {
  /** A map of feature name to the bundle path of the feature */
  bundlePaths: {
    [key: string]: string;
  };
  /** A map of project roots to their paths */
  projectFilepaths: {
    /** The folder path where the template files live */
    templatesRoot: string;
    /** The folder path where the compiled files live */
    distRoot: string;
    /** The folder path where the compiled hydration bundles live */
    hydrationBundleOutputRoot: string;
    /** The folder path where the compiled server bundles live */
    serverBundleOutputRoot: string;
  };
  /** If the bundler used generates a manifest.json then this field will contain that json object */
  bundlerManifest?: any;
};

/**
 * The shape of the data passed directly to the different template functions with the
 * exception of the render function (getPath, getHeadConfig, etc).
 *
 * @public
 */
export interface TemplateProps {
  /** The entire document returned after applying the stream to a single entity */
  document: Record<string, any>;
  /** Additional metadata added by the toolchain */
  __meta: {
    /** Specifies if the data is returned in development or production mode */
    mode: "development" | "production";
    /** A manifest of bundled files present during production mode */
    manifest?: Manifest;
  };
}

/**
 * The shape of the data passed directly to the template's render function.
 * Extends the {@link TemplateProps} interface and has the additions of a path
 * and a relativePrefixToRoot field.
 *
 * @public
 */
export interface TemplateRenderProps extends TemplateProps {
  /**
   * The path that the generated file will live at on the site, as defined
   * by the {@link GetPath} function.
   */
  path: string;
  /**
   * The relative path from the generated page to the root of the site.
   * i.e. The path example/path/foo would have the relativePrefixToRoot
   * of '../../'.
   */
  relativePrefixToRoot: string;
}